+++ test/ProofOfCapital/ProofOfCapital.initialization.t.sol
@@ -71,7 +71,7 @@ contract ProofOfCapitalInitializationTest is BaseTest {
     // Test MultiplierTooHigh error
     function testInitializeMultiplierTooHigh() public {
         ProofOfCapital.InitParams memory params = getValidParams();
-        params.levelDecreaseMultiplierafterTrend = Constants.PERCENTAGE_DIVISOR; // Invalid: equal to divisor
+        params.levelDecreaseMultiplierafterTrend = int256(Constants.PERCENTAGE_DIVISOR); // Invalid: equal to divisor
 
         vm.expectRevert(ProofOfCapital.MultiplierTooHigh.selector);
         new ProofOfCapital(params);
@@ -79,7 +79,7 @@ contract ProofOfCapitalInitializationTest is BaseTest {
 
     function testInitializeMultiplierTooHighAboveDivisor() public {
         ProofOfCapital.InitParams memory params = getValidParams();
-        params.levelDecreaseMultiplierafterTrend = Constants.PERCENTAGE_DIVISOR + 1; // Invalid: above divisor
+        params.levelDecreaseMultiplierafterTrend = int256(Constants.PERCENTAGE_DIVISOR + 1); // Invalid: above divisor
 
         vm.expectRevert(ProofOfCapital.MultiplierTooHigh.selector);
         new ProofOfCapital(params);
@@ -87,14 +87,14 @@ contract ProofOfCapitalInitializationTest is BaseTest {
 
     function testInitializeMultiplierValidAtBoundary() public {
         ProofOfCapital.InitParams memory params = getValidParams();
-        params.levelDecreaseMultiplierafterTrend = Constants.PERCENTAGE_DIVISOR - 1; // Valid: just below divisor
+        params.levelDecreaseMultiplierafterTrend = int256(Constants.PERCENTAGE_DIVISOR - 1); // Valid: just below divisor
         params.offsetTokens = 100e18; // Smaller offset to avoid overflow in calculations
 
         // Should not revert
         ProofOfCapital proofOfCapital = new ProofOfCapital(params);
 
         // Verify multiplier was set
-        assertEq(proofOfCapital.levelDecreaseMultiplierafterTrend(), Constants.PERCENTAGE_DIVISOR - 1);
+        assertEq(proofOfCapital.levelDecreaseMultiplierafterTrend(), int256(Constants.PERCENTAGE_DIVISOR - 1));
     }
 
     // Test MultiplierTooLow error for levelIncreaseMultiplier
